Fix patch

+++ framework/commands/uninstall.rb
@@ -27,9 +27,11 @@ EOS
       else
         CLI.notice "Uninstall package #{CLI.green package.name} #{CLI.blue package.version} ..."
         PackageLinker.unlink package
-        FileUtils.rm_rf package.prefix if Dir.exist? package.prefix
-        # Remove empty directory.
-        FileUtils.rmdir File.dirname(package.prefix) if Dir.glob("#{File.dirname(package.prefix)}/*").size == 0
+        if Dir.exist? package.prefix
+          FileUtils.rm_rf package.prefix
+          # Remove empty directory.
+          FileUtils.rmdir File.dirname(package.prefix) if Dir.glob("#{File.dirname(package.prefix)}/*").size == 0
+        end
         History.remove_install package
       end
     end
